fix(ci): python-coverage-shards — split oversized test_workspaces_part_002 under line limit
test_first_party_code_files_stay_under_line_limit failed because a 2-line
addition (404b1ac85) pushed test_workspaces_part_002.py to 1501 lines, one
over the 1500 MAX_FIRST_PARTY_FILE_LINES guardrail. Move the two trailing
durable idempotency-replay tests into a new test_workspaces_part_008.py
(with the helpers they need), bringing part_002 to 1367 lines.
