+++ tests/unit/api/test_workspaces_parts/test_workspaces_part_002.py
@@ -1365,137 +1365,3 @@ class TestCreateWorkspacePart001:
         assert matched is True
         assert lock.enters == 2
         assert lock.exits == 2
-
-    @pytest.mark.unit
-    async def test_v1_cache_hash_conflict_uses_durable_replay_before_conflict(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-    ) -> None:
-        request = _request_with_disk_check()
-        payload = WorkspaceCreateRequest.model_validate(
-            {**_MINIMAL_BODY, "task_title": "durable replay after stale cache hash"}
-        )
-        idempotency_key = "workspace-v1-stale-cache-hash"
-        replay_key_cache = workspaces_route._workspace_create_idempotency_replay_key_cache(  # noqa: SLF001
-            request
-        )
-        replay_key_cache.remember_hash(
-            idempotency_key=idempotency_key,
-            request_hash="stale-cache-hash",
-        )
-        existing = SimpleNamespace(
-            id="ws_v1_stale_hash_replay",
-            status=WorkspaceStatus.requested.value,
-            version=3,
-            created_at=datetime(2026, 5, 15, tzinfo=UTC),
-            repo_url=payload.repo_url,
-            branch_base=payload.branch_base,
-            task_tag=payload.task.task_tag,
-            task_title=payload.task_title,
-            task_prompt=payload.task_prompt,
-            task_external_id=payload.task_external_id,
-            agent=payload.agent.value,
-            env_profile=payload.env_profile,
-            test_commands=list(payload.test_commands),
-            requires_database=payload.requires_database,
-            task_attempt=None,
-        )
-        lock_keys: list[str] = []
-        lookup_keys: list[str] = []
-        create_calls: list[str | None] = []
-
-        async def tracked_lock(_self: WorkspaceRepository, key: str) -> None:
-            lock_keys.append(key)
-
-        async def tracked_lookup(_self: WorkspaceRepository, key: str) -> object:
-            lookup_keys.append(key)
-            return existing
-
-        async def fail_create(_self: WorkspaceRepository, **kwargs: object) -> None:
-            create_calls.append(kwargs.get("idempotency_key"))
-            raise AssertionError("durable replay must not create a new workspace")
-
-        monkeypatch.setattr(WorkspaceRepository, "acquire_idempotency_key_lock", tracked_lock)
-        monkeypatch.setattr(WorkspaceRepository, "get_by_idempotency_key", tracked_lookup)
-        monkeypatch.setattr(WorkspaceRepository, "create", fail_create)
-
-        response = await workspaces_route.create_workspace(
-            payload,
-            request=request,  # type: ignore[arg-type]
-            idempotency_key=idempotency_key,
-            settings=_workspace_request_admission_settings(limit=10),
-            session=SimpleNamespace(info={}, bind=None),  # type: ignore[arg-type]
-        )
-
-        assert not isinstance(response, JSONResponse)
-        assert response.workspace_id == existing.id
-        assert lock_keys == [idempotency_key]
-        assert lookup_keys == [idempotency_key]
-        assert create_calls == []
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize(
-        ("payload", "idempotency_key"),
-        [
-            pytest.param(
-                WorkspaceCreateRequest.model_validate(
-                    {**_MINIMAL_BODY, "task_title": "known missing replay v1"}
-                ),
-                "known-missing-workspace-v1",
-                id="v1",
-            ),
-            pytest.param(
-                WorkspaceCreateRequest.model_validate(_v2_body(title="known missing replay v2")),
-                "known-missing-workspace-v2",
-                id="v2",
-            ),
-        ],
-    )
-    async def test_known_replay_key_db_miss_returns_conflict_without_create(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-        payload: WorkspaceCreateRequest,
-        idempotency_key: str,
-    ) -> None:
-        request = _request_with_disk_check()
-        replay_key_cache = workspaces_route._workspace_create_idempotency_replay_key_cache(  # noqa: SLF001
-            request
-        )
-        replay_key_cache.remember(
-            payload,
-            idempotency_key=idempotency_key,
-            api_version=workspaces_route._WORKSPACE_CREATE_API_VERSION,  # noqa: SLF001
-        )
-        lock_keys: list[str] = []
-        lookup_keys: list[str] = []
-        create_calls: list[str | None] = []
-
-        async def tracked_lock(_self: WorkspaceRepository, key: str) -> None:
-            lock_keys.append(key)
-
-        async def tracked_lookup(_self: WorkspaceRepository, key: str) -> None:
-            lookup_keys.append(key)
-
-        async def fail_create(*_args: object, **kwargs: object) -> None:
-            create_calls.append(kwargs.get("idempotency_key"))
-            raise AssertionError("known replay-key durable miss must not create a workspace")
-
-        monkeypatch.setattr(WorkspaceRepository, "acquire_idempotency_key_lock", tracked_lock)
-        monkeypatch.setattr(WorkspaceRepository, "get_by_idempotency_key", tracked_lookup)
-        monkeypatch.setattr(workspaces_route, "create_workspace_row_checked", fail_create)
-
-        session = SimpleNamespace(info={}, bind=None)
-        response = await workspaces_route.create_workspace(
-            payload,
-            request=request,  # type: ignore[arg-type]
-            idempotency_key=idempotency_key,
-            settings=_workspace_request_admission_settings(limit=10),
-            session=session,  # type: ignore[arg-type]
-        )
-
-        assert isinstance(response, JSONResponse)
-        assert response.status_code == 409
-        assert json.loads(response.body)["error_code"] == "IDEMPOTENCY_REPLAY_UNAVAILABLE"
-        assert lock_keys == [idempotency_key]
-        assert lookup_keys == [idempotency_key]
-        assert create_calls == []
